+++ thotlib/dialogue/printmenu.c
@@ -357,7 +357,7 @@ static void Print (char *name, char *dir, char *thotSch, char *thotDoc,
        printArgc++;
 #else /* _WINDOWS */
 #ifndef _WX
-       sprintf (&cmd[i], " -w%u", (unsigned int) FrRef[0]);
+       sprintf (&cmd[i], " -w%lu", (intptr_t) FrRef[0]);
 #else /* _WX */
        sprintf (&cmd[i], " -w%u", (unsigned int) FrameTable[0].WdFrame);
 #endif /* _WX */
